Set the infos cache of the keyword after a url is added/updated/deleted (#4008)
Set the infos cache of the keyword after a url is added/updated/deleted to/from the database.

This will allow it to overwrite the cache with the updated keyword if a cache was already set on the request.

This will prevent plugins accessing cached data from accessing stale data.

Adds a new function update_infos_if_exists to the YDB class to update values on an existing cache if it exists

Add tests to test the logic.

Fixes #3772

Co-authored-by: ྅༻ Ǭɀħ ༄༆ཉ <ozh@ozh.org>